team_CakeConstructor/src/main/java/lv/javaguru/java2/cakeConstructor/newApp/core/services/GetAllIngredientsService.java
package lv.javaguru.java2.cakeConstructor.newApp.core.services;

import lv.javaguru.java2.cakeConstructor.newApp.core.database.jpa.JpaIngredientRepository;
import lv.javaguru.java2.cakeConstructor.newApp.core.domain.Ingredient;
import lv.javaguru.java2.cakeConstructor.newApp.core.requests.GetAllIngredientsRequest;
import lv.javaguru.java2.cakeConstructor.newApp.core.response.GetAllIngredientsResponse;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Component
@Transactional
public class GetAllIngredientsService {

    @Autowired private JpaIngredientRepository ingredientRepository;


    public GetAllIngredientsResponse execute(GetAllIngredientsRequest request){
        List<Ingredient> ingredients = ingredientRepository.findAll();
        return new GetAllIngredientsResponse(ingredients);
    }
}
